Change the functionality of this role
There is no need to download the entire cisagov/code-gov-update
repository to the host when all we care about is the Docker Compose
configurations. Instead we can store a deployment specific Compose
file in this role and copy that to the host. This is, in my opinion, a
better model to follow so that the cisagov/code-gov-update project can
focus on managing the Docker image only.