Store the Docker Compose configuration in this project
This moves the Docker Compose configuration out of a
`docker-compose.override.yml` file stored in the configuration for the
cisagov/cide-gov-update image and into this Ansible role. This
ensures that the necessary configuration to deploy this Docker image is
stored with the Ansible role and changes to the configuration will be
limited to this repository. With this the image configuration can
change and until the Docker configuration in this role is modified it
will not impact this role's usage.